requirements.txt
asgiref==3.10.0
Django==5.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-jazzmin==3.0.1
django-taggit==6.1.0
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
